requirements.txt
pillow==10.4.0
streamlit==1.40.1
